backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.0
boto3==1.35.0
geopandas==1.0.1
shapely==2.0.6
